add retry & timeout logic to coordinator client

+++ prover_rust/src/config.rs
@@ -13,9 +13,9 @@ pub struct CircuitConfig {
 #[derive(Debug, Serialize, Deserialize)]
 pub struct CoordinatorConfig {
     pub base_url: String,
-    pub retry_count: u16,
-    pub retry_wait_time_sec: u32,
-    pub connection_timeout_sec: u32,
+    pub retry_count: u32,
+    pub retry_wait_time_sec: u64,
+    pub connection_timeout_sec: u64,
 }
 
 #[derive(Debug, Serialize, Deserialize)]
